requirements.txt
pyTelegramBotAPI>=4.12.0
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
supabase>=2.0.0
pytz>=2023.3
schedule>=1.2.0
